Move hipclang funcs into corresponding headers
